pkg/operator/encryption/kms: move checker from preflight to parent package
Relocate checker.go and checker_test.go into the kms package and export the
type, constructor, and check method (Checker / NewChecker / Check) so the
preflight command consumes it as a parent-package primitive. The status
poll helpers, struct fields, and healthzOK stay unexported; the in-package
test keeps reaching them directly.

preflight's cmd_e2e_test.go is unchanged and still green, confirming the
move is a behavior-preserving refactor of the preflight package.
